requirements.txt
torch>=2.1.0
numpy>=1.24
scipy>=1.11
pandas>=2.0
scikit-learn>=1.3
pywt>=1.5
PyYAML>=6.0
matplotlib>=3.8
tensorboard>=2.15
tqdm>=4.66
